change the code of sp2gcl
